Move curves factory
