+++ src/dgcv/curves/curves.py
@@ -15,8 +15,8 @@ from dgcv.configuration.cfg import config
 from dgcv.core.execution_parameters import Parameters
 from dgcv.core.validator import Disconnection_Model
 from dgcv.curves.importer import CurvesImporter
-from dgcv.curves.producer import ProducerCurves, get_cfg_oc_name
 from dgcv.files import manage_files
+from dgcv.model.curves import ProducerCurves, get_cfg_oc_name
 from dgcv.model.parameters import Gen_init, Gen_params
