+++ src/dgcv/model/curves.py
