+++ tests/dgcv/validation/test_sanity_checks.py
@@ -143,7 +143,7 @@ def test_internal_lines():
     )
 
 
-@pytest.mark.skipif(shutil.which("dynawo.sh"), reason="Dynawo installed")
+@pytest.mark.skipif(shutil.which("dynawo.sh") is not None, reason="Dynawo installed")
 def test_launchers():
     with pytest.raises(OSError) as pytest_wrapped_e:
         sanity_checks.check_launchers("dynawo.sh")
